Update README.md
Now that I am jamming on Git and GitHub, I would like to update our project's README.md file to add some other helpful information in a central location that we can easily find! Review the changes and if everyone is okay with them, the last person to review it should go ahead and do the merge to main! ;-)